requirements.txt
# FastAPI:
fastapi[standard]
uvicorn

# Authentication:
pyjwt
passlib[bcrypt]
pyotp
qrcode

# Database:
sqlalchemy

# Testing
pytest
httpx
